# perf: optimize client builds