--- a/packages/brisa/src/utils/client-build-plugin/process-client-ast/index.ts
+++ b/packages/brisa/src/utils/client-build-plugin/process-client-ast/index.ts
@@ -1,19 +1,49 @@
 import type { ESTree } from 'meriyah';
+import { join } from 'node:path';
 import { logWarning } from '@/utils/log/log-build';
-import AST from '@/utils/ast';
 import { toInline } from '@/helpers';
+import AST from '@/utils/ast';
 
 const { generateCodeFromAST } = AST('tsx');
+const brisaClientPath = join('brisa', 'client', 'index.js');
 
 export default function processClientAst(ast: ESTree.Program, path = '') {
   let i18nKeys = new Set<string>();
   let useI18n = false;
   const logs: any[] = [];
+  const isBrisaClient = path.endsWith(brisaClientPath);
   let isDynamicKeysSpecified = false;
 
   const newAst = JSON.parse(JSON.stringify(ast), (key, value) => {
-    useI18n ||= value?.type === 'Identifier' && value?.name === 'i18n';
+    useI18n ||=
+      !isBrisaClient && value?.type === 'Identifier' && value?.name === 'i18n';
 
+    if (
+      value?.type === 'CallExpression' &&
+      ((value?.callee?.type === 'Identifier' && value?.callee?.name === 't') ||
+        (value?.callee?.property?.type === 'Identifier' &&
+          value?.callee?.property?.name === 't'))
+    ) {
+      if (value?.arguments?.[0]?.type === 'Literal') {
+        i18nKeys.add(value?.arguments?.[0]?.value);
+      } else {
+        logs.push(value);
+      }
+    }
+
+    // Add dynamic keys from: MyWebComponent.i18nKeys = ['footer', /projects.*title/];
+    if (
+      value?.type === 'ExpressionStatement' &&
+      value.expression.left?.property?.name === 'i18nKeys' &&
+      value.expression?.right?.type === 'ArrayExpression'
+    ) {
+      for (const element of value.expression.right.elements ?? []) {
+        i18nKeys.add(element.value);
+        isDynamicKeysSpecified = true;
+      }
+      // Remove the expression statement
+      return null;
+    }
     // Remove react/jsx-runtime import, some transpilers like @swc add it,
     // but we are not using jsx-runtime here, we are using jsx-buildtime
     if (
@@ -48,35 +78,8 @@ export default function processClientAst(ast: ESTree.Program, path = '') {
       return null;
     }
 
-    if (
-      value?.type === 'CallExpression' &&
-      ((value?.callee?.type === 'Identifier' && value?.callee?.name === 't') ||
-        (value?.callee?.property?.type === 'Identifier' &&
-          value?.callee?.property?.name === 't'))
-    ) {
-      if (value?.arguments?.[0]?.type === 'Literal') {
-        i18nKeys.add(value?.arguments?.[0]?.value);
-      } else {
-        logs.push(value);
-      }
-    }
-
-    // Add dynamic keys from: MyWebComponent.i18nKeys = ['footer', /projects.*title/];
-    if (
-      value?.type === 'ExpressionStatement' &&
-      value.expression.left?.property?.name === 'i18nKeys' &&
-      value.expression?.right?.type === 'ArrayExpression'
-    ) {
-      for (const element of value.expression.right.elements ?? []) {
-        i18nKeys.add(element.value);
-        isDynamicKeysSpecified = true;
-      }
-      // Remove the expression statement
-      return null;
-    }
-
-    // Remove arrays with empty values
-    if (Array.isArray(value)) return value.filter((v) => v);
+    // Clean null values inside arrays
+    if (Array.isArray(value)) return value.filter(Boolean);
 
     return value;
   });
@@ -107,7 +110,65 @@ export default function processClientAst(ast: ESTree.Program, path = '') {
     );
   }
 
-  if (!useI18n) i18nKeys = new Set();
+  // This is a workaround to in a post-analysis collect all i18nKeys and useI18n from the
+  // entrypoint to inject the i18n bridge and clean these variables. That is, they are not
+  // real gobal variables, it is a communication between dependency graph for the post-analysis.
+  //
+  // It is necessary to think that each file can be connected to different entrypoints, so at
+  // this point we note the keys and if it uses i18n (lang or other attributes without translations).
+  //
+  // Communication variables: window.i18nKeys & window.useI18n
+  if (useI18n && i18nKeys.size) {
+    newAst.body.push({
+      type: 'ExpressionStatement',
+      expression: {
+        type: 'AssignmentExpression',
+        operator: '=',
+        left: {
+          type: 'MemberExpression',
+          object: {
+            type: 'Identifier',
+            name: 'window',
+          },
+          property: {
+            type: 'Identifier',
+            name: 'i18nKeys',
+          },
+        },
+        right: {
+          type: 'ArrayExpression',
+          elements: Array.from(i18nKeys).map((key) => ({
+            type: 'Literal',
+            value: key,
+          })),
+        },
+      },
+    });
+  }
+  if (useI18n) {
+    newAst.body.push({
+      type: 'ExpressionStatement',
+      expression: {
+        type: 'AssignmentExpression',
+        operator: '=',
+        left: {
+          type: 'MemberExpression',
+          object: {
+            type: 'Identifier',
+            name: 'window',
+          },
+          property: {
+            type: 'Identifier',
+            name: 'useI18n',
+          },
+        },
+        right: {
+          type: 'Literal',
+          value: useI18n,
+        },
+      },
+    });
+  }
 
-  return { useI18n, i18nKeys, ast: newAst };
+  return newAst;
 }