--- a/packages/brisa/src/utils/client-build/generate-entrypoint-code/index.ts
+++ b/packages/brisa/src/utils/client-build/generate-entrypoint-code/index.ts
@@ -10,7 +10,7 @@ import { getConstants } from '@/constants';
 import { normalizePath } from '../normalize-path';
 import snakeToCamelCase from '@/utils/snake-to-camelcase';
 
-type EntrpypointOptions = {
+type EntrypointOptions = {
   webComponentsList: Record<string, string>;
   useContextProvider: boolean;
   integrationsPath?: string | null;
@@ -20,81 +20,92 @@ export async function generateEntryPointCode({
   webComponentsList,
   useContextProvider,
   integrationsPath,
-}: EntrpypointOptions) {
+}: EntrypointOptions) {
   const { IS_DEVELOPMENT } = getConstants();
-  let useWebContextPlugins = false;
   const entries = Object.entries(webComponentsList);
 
-  if (!useContextProvider && !entries.length) {
-    return {
-      code: '',
-      useWebContextPlugins: false,
-    };
+  if (!useContextProvider && entries.length === 0) {
+    return { code: '', useWebContextPlugins: false };
   }
 
-  // Note: JS imports in Windows have / instead of \, so we need to replace it
-  // Note: Using "require" for component dependencies not move the execution
-  // on top avoiding missing global variables as window._P (P = plugins)
-  let imports = entries
-    .map(([name, path]) =>
-      path[0] === '{'
-        ? `require("${normalizePath(path)}");`
-        : `import ${snakeToCamelCase(name)} from "${path.replaceAll(sep, '/')}";`,
-    )
-    .join('\n');
-
-  // Add web context plugins import only if there is a web context plugin
-  if (integrationsPath) {
-    const module = await import(integrationsPath);
-    if (module.webContextPlugins?.length > 0) {
-      useWebContextPlugins = true;
-      imports += `import {webContextPlugins} from "${integrationsPath}";`;
-    }
-  }
-
-  const defineElement =
-    'const defineElement = (name, component) => name && !customElements.get(name) && customElements.define(name, component);';
+  const { imports, useWebContextPlugins } = await getImports(
+    entries,
+    integrationsPath,
+  );
 
   const customElementKeys = entries
     .filter(([_, path]) => path[0] !== '{')
-    .map(([k]) => k);
+    .map(([key]) => key);
 
-  if (useContextProvider) {
-    customElementKeys.unshift('context-provider');
-  }
-
-  if (IS_DEVELOPMENT) {
-    customElementKeys.unshift('brisa-error-dialog');
-  }
-
-  const customElementsDefinitions = customElementKeys
-    .map((k) => `defineElement("${k}", ${snakeToCamelCase(k)});`)
-    .join('\n');
+  addOptionalComponents(customElementKeys, useContextProvider, IS_DEVELOPMENT);
 
   let code = '';
 
   if (useContextProvider) {
-    const contextProviderCode =
-      injectClientContextProviderCode() as unknown as string;
-    code += contextProviderCode;
+    code += injectClientContextProviderCode();
   }
 
-  // IS_DEVELOPMENT to avoid PROD and TEST environments
   if (IS_DEVELOPMENT) {
-    const brisaDialogErrorCode = (await injectBrisaDialogErrorCode()).replace(
-      '__FILTER_DEV_RUNTIME_ERRORS__',
-      getFilterDevRuntimeErrors(),
-    );
-    code += brisaDialogErrorCode;
+    code += await getDevelopmentCode();
   }
 
-  // Inject web context plugins to window to be used inside web components
   if (useWebContextPlugins) {
     code += 'window._P=webContextPlugins;\n';
   }
 
-  code += `${imports}\n`;
-  code += `${defineElement}\n${customElementsDefinitions}`;
+  code += `${imports}\n${getDefineElementCode(customElementKeys)}`;
 
   return { code, useWebContextPlugins };
+}
+
+async function getImports(
+  entries: [string, string][],
+  integrationsPath?: string | null,
+) {
+  const imports = entries.map(([name, path]) =>
+    path[0] === '{'
+      ? `require("${normalizePath(path)}");`
+      : `import ${snakeToCamelCase(name)} from "${path.replaceAll(sep, '/')}";`,
+  );
+
+  if (integrationsPath) {
+    const module = await import(integrationsPath);
+    if (module.webContextPlugins?.length > 0) {
+      imports.push(`import {webContextPlugins} from "${integrationsPath}";`);
+      return { imports: imports.join('\n'), useWebContextPlugins: true };
+    }
+  }
+
+  return { imports: imports.join('\n'), useWebContextPlugins: false };
+}
+
+function getDefineElementCode(keys: string[]): string {
+  const defineElementCode =
+    'const defineElement = (name, component) => name && !customElements.get(name) && customElements.define(name, component);';
+  const definitions = keys
+    .map((key) => `defineElement("${key}", ${snakeToCamelCase(key)});`)
+    .join('\n');
+
+  return `${defineElementCode}\n${definitions}`;
+}
+
+async function getDevelopmentCode(): Promise<string> {
+  const brisaDialogErrorCode = (await injectBrisaDialogErrorCode()).replace(
+    '__FILTER_DEV_RUNTIME_ERRORS__',
+    getFilterDevRuntimeErrors(),
+  );
+  return brisaDialogErrorCode;
+}
+
+function addOptionalComponents(
+  keys: string[],
+  useContextProvider: boolean,
+  isDevelopment: boolean,
+) {
+  if (useContextProvider) {
+    keys.unshift('context-provider');
+  }
+  if (isDevelopment) {
+    keys.unshift('brisa-error-dialog');
+  }
 }