--- a/packages/brisa/src/utils/client-build/generate-entrypoint-code/index.ts
+++ b/packages/brisa/src/utils/client-build/generate-entrypoint-code/index.ts
@@ -77,6 +77,7 @@ async function getImports(
   );
 
   if (integrationsPath) {
+    // TODO: cache dynamic import can improve performance? We need to try it
     const module = await import(integrationsPath);
 
     if (module.webContextPlugins?.length > 0) {