… docs/building-your-application/configuring/plugins.md
… packages/brisa/src/core/test/run-web-components/index.ts
… packages/brisa/src/utils/client-build-plugin/add-i18n-bridge/index.test.ts
… packages/brisa/src/utils/client-build-plugin/add-i18n-bridge/index.ts
… packages/brisa/src/utils/client-build-plugin/index.test.ts
… packages/brisa/src/utils/client-build-plugin/integration.test.tsx
… packages/brisa/src/utils/client-build-plugin/process-client-ast/index.test.ts
… packages/brisa/src/utils/client-build/fs-temp-entrypoint-manager/index.test.ts
… packages/brisa/src/utils/client-build/generate-entrypoint-code/index.test.ts
… packages/brisa/src/utils/client-build/get-client-build-details/index.test.ts
… packages/brisa/src/utils/client-build/get-client-build-details/index.ts
… packages/brisa/src/utils/client-build/get-temp-page-name/index.test.ts
… packages/brisa/src/utils/client-build/get-temp-page-name/index.ts
… packages/brisa/src/utils/client-build/index.ts
… packages/brisa/src/utils/client-build/layout-build/index.test.ts
… packages/brisa/src/utils/client-build/layout-build/index.ts
… packages/brisa/src/utils/client-build/normalize-path/index.test.ts
… packages/brisa/src/utils/client-build/normalize-path/index.ts
… packages/brisa/src/utils/client-build/pages-build/index.test.ts
… packages/brisa/src/utils/client-build/pages-build/index.ts
… packages/brisa/src/utils/client-build/pre-entrypoint-analysis/index.test.ts
… packages/brisa/src/utils/client-build/pre-entrypoint-analysis/index.ts
… packages/brisa/src/utils/client-build/process-i18n/index.test.ts
… packages/brisa/src/utils/client-build/process-i18n/index.ts
… packages/brisa/src/utils/client-build/process-i18n/inject-bridge.ts
… packages/brisa/src/utils/client-build/run-build/index.test.ts
… packages/brisa/src/utils/client-build/run-build/index.ts
… packages/brisa/src/utils/client-build/types.ts
… packages/brisa/src/utils/compile-files/index.test.ts
… packages/brisa/src/utils/compile-files/index.ts