Related https://github.com/brisa-build/brisa/issues/162
Related https://github.com/brisa-build/brisa/issues/556

Finally the PR is ready 🔥 

The improvement is not as much as I thought, but it is a good step. Good parts:

- We already have IIFE implemented with Bun
- Instead of using multi-build we use only 1 for the client.

Things to look at if we can look at how to improve:

I noticed that currently what subtracts the most milliseconds are the barrel files we create for the Bun builder to have. Currently this is necessary because Bun needs the files to do the build and these files do not exist previously. To understand the actual process; first we build the server and in this step we analyze which Web Components each page needs (or not). Then for all the pages that have Web Components a temporary file is generated with the imports of the web components so that the client build takes them into account as entrypoints and resolves the dependencies and generates the files that the client pages need.

Only if we can avoid this step in the future and create the barrel files directly for use in runtime without the step of going through fopen, then we will improve about 40% of the current time.

## Results

**hardware used:** 23.6.0 arm64 arm
**note:** all pages have server and client code  (because only server without client code are faster)

![build-new](https://github.com/user-attachments/assets/ee386aca-e7b9-4104-bcf2-65927989600e)


**Before:**

- 10 pages - 359.07ms
- 20 pages - 851.17ms
- 30 pages - 1553.64ms
- 50 pages - 2405.21ms
- 80 pages - 4001.66ms
- 100 pages - 5059.29ms
- 150 pages - 7578.01ms
- 300 pages - 14848.46ms
- 400 pages - 20620.80ms
- 600 pages - 34038.54ms
- 1000 pages - 68510.37ms

**Now:**

- 10 pages - 335.64ms
- 20 pages - 723.02ms
- 30 pages - 1291.97ms
- 50 pages - 1866.04ms
- 80 pages - 3262.58ms
- 100 pages - 4406.63ms
- 150 pages - 5782.00ms
- 300 pages - 11660.03ms
- 400 pages - 17336.42ms
- 600 pages - 25952.91ms
- 1000 pages - 45488.65ms